python_langchain/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.14
    # via -r requirements.in
aiosignal==1.3.2
    # via aiohttp
anyio==4.8.0
    # via httpx
attrs==25.3.0
    # via aiohttp
awscrt==0.23.8
    # via botocore
azure-core==1.32.0
    # via
    #   azure-cosmos
    #   azure-identity
azure-cosmos==4.9.0
    # via -r requirements.in
azure-identity==1.13.0
    # via -r requirements.in
black==22.12.0
    # via -r requirements.in
boto3[crt]==1.37.13
    # via -r requirements.in
botocore[crt]==1.37.13
    # via
    #   boto3
    #   s3transfer
certifi==2020.12.5
    # via
    #   -r requirements.in
    #   httpcore
    #   httpx
    #   requests
cffi==1.14.5
    # via cryptography
charset-normalizer==3.2.0
    # via requests
click==8.1.8
    # via black
coverage[toml]==7.7.0
    # via pytest-cov
cryptography==3.4.6
    # via
    #   azure-identity
    #   msal
    #   pyjwt
dnspython==2.3.0
    # via -r requirements.in
docopt==0.6.2
    # via -r requirements.in
duckdb==1.2.1
    # via -r requirements.in
faker==37.0.0
    # via -r requirements.in
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via -r requirements.in
idna==2.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
jinja2==2.11.3
    # via -r requirements.in
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
markupsafe==1.1.1
    # via jinja2
msal==1.22.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.0.0
    # via azure-identity
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via black
packaging==20.9
    # via pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.6
    # via black
pluggy==1.5.0
    # via pytest
portalocker==2.7.0
    # via msal-extensions
propcache==0.3.0
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via -r requirements.in
pycparser==2.20
    # via cffi
pyjwt[crypto]==2.7.0
    # via
    #   msal
    #   pyjwt
pymongo==3.11.3
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pytest==8.3.5
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.25.3
    # via -r requirements.in
pytest-cov==6.0.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2025.1
    # via -r requirements.in
redis==5.2.1
    # via -r requirements.in
requests==2.31.0
    # via
    #   azure-core
    #   msal
s3transfer==0.11.4
    # via boto3
six==1.17.0
    # via
    #   -r requirements.in
    #   azure-core
    #   azure-identity
    #   python-dateutil
sniffio==1.3.1
    # via anyio
typing-extensions==4.6.3
    # via
    #   azure-core
    #   azure-cosmos
tzdata==2025.1
    # via faker
urllib3==1.26.20
    # via
    #   botocore
    #   requests
yarl==1.18.3
    # via aiohttp
